+++ workspace/Avalon/src/campaigns/SpyingCampaign.java
@@ -4,7 +4,7 @@ import departments.Department;
 
 public class SpyingCampaign extends ResearchCampaign {
 
-	public SpyingCampaign(Department department, String title, int cost,
+	public SpyingCampaign(Department department, String title, double cost,
 			int duration, int successProbability, int level, String description) {
 		super(department, title, cost, duration, successProbability, level,
 				description);
